requirements.txt
accelerate==0.29.1
certifi==2024.2.2
charset-normalizer==3.3.2
clip==0.2.0
cmake==3.29.0.1
diffusers==0.20.0
einops==0.7.0
filelock==3.13.3
fsspec==2024.3.1
huggingface-hub==0.22.2
idna==3.6
importlib_metadata==7.1.0
Jinja2==3.1.3
lit==18.1.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==24.0
pillow==10.3.0
psutil==5.9.8
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
sympy==1.12
tokenizers==0.15.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.2
transformers==4.39.3
triton==2.0.0
typing_extensions==4.11.0
urllib3==2.2.1
zipp==3.18.1
